+++ opencti-platform/opencti-front/src/private/components/profile/ForcePasswordChange.jsx
@@ -1,14 +1,10 @@
 import React from 'react';
-import { Formik, Form, Field } from 'formik';
-import * as Yup from 'yup';
-import { graphql } from 'react-relay';
 import { Box, Stack, useTheme } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
-import Button from '@common/button/Button';
-import TextField from '../../../components/TextField';
 import PasswordPolicies from '../common/form/PasswordPolicies';
 import { useFormatter } from '../../../components/i18n';
-import { commitMutation, handleErrorInForm, MESSAGING$ } from '../../../relay/environment';
+import { MESSAGING$ } from '../../../relay/environment';
+import ForcePasswordChangeForm from '../../../components/ForcePasswordChangeForm';
 import logoDark from '../../../static/images/logo_text_dark.png';
 import logoLight from '../../../static/images/logo_text_light.png';
 import logoFiligranBaselineDark from '../../../static/images/logo_filigran_baseline_dark.svg';
@@ -16,46 +12,14 @@ import logoFiligranGradientDark from '../../../static/images/logo_filigran_gradi
 import logoFiligranBaselineLight from '../../../static/images/logo_filigran_baseline_light.svg';
 import logoFiligranGradientLight from '../../../static/images/logo_filigran_gradient_light.svg';
 
-const forcePasswordChangeMutation = graphql`
-  mutation ForcePasswordChangeMutation(
-    $input: [EditInput]!
-  ) {
-    meEdit(input: $input) {
-      id
-      password_valid_until
-    }
-  }
-`;
-
-const passwordValidation = (t) => Yup.object().shape({
-  password: Yup.string().required(t('This field is required')),
-  confirmation: Yup.string()
-    .oneOf([Yup.ref('password'), null], t('The values do not match'))
-    .required(t('This field is required')),
-});
-
 const ForcePasswordChange = () => {
   const theme = useTheme();
   const navigate = useNavigate();
   const { t_i18n } = useFormatter();
 
-  const onSubmit = (values, { setSubmitting, resetForm, setErrors }) => {
-    commitMutation({
-      mutation: forcePasswordChangeMutation,
-      variables: {
-        input: [{ key: 'password', value: [values.password] }],
-      },
-      onCompleted: () => {
-        setSubmitting(false);
-        MESSAGING$.notifySuccess(t_i18n('The password has been updated'));
-        resetForm();
-        navigate('/dashboard', { replace: true });
-      },
-      onError: (error) => {
-        handleErrorInForm(error, setErrors);
-        setSubmitting(false);
-      },
-    });
+  const handleSuccess = () => {
+    MESSAGING$.notifySuccess(t_i18n('The password has been updated'));
+    navigate('/dashboard', { replace: true });
   };
 
   return (
@@ -79,66 +43,22 @@ const ForcePasswordChange = () => {
           width={180}
         />
         <Stack gap={1} sx={{ width: 500 }}>
-          <Formik
-            enableReinitialize={true}
-            validateOnMount={true}
-            initialValues={{ password: '', confirmation: '' }}
-            validationSchema={passwordValidation(t_i18n)}
-            onSubmit={onSubmit}
+          <Box
+            sx={{
+              background: theme.palette.background.paper,
+              borderRadius: 1,
+              padding: theme.spacing(3),
+            }}
           >
-            {({ isSubmitting, isValid, values }) => (
-              <Form
-                style={{
-                  flex: 1,
-                  display: 'flex',
-                  flexDirection: 'column',
-                  width: '100%',
-                  height: '100%',
-                }}
-              >
-                <Box
-                  sx={{
-                    background: theme.palette.background.paper,
-                    borderRadius: 1,
-                    padding: theme.spacing(3),
-                  }}
-                >
-                  <Box sx={{ width: '100%', mt: 2 }}>
-                    <PasswordPolicies value={values.password} />
-                  </Box>
-                  <Field
-                    component={TextField}
-                    name="password"
-                    label={t_i18n('New password')}
-                    type="password"
-                    fullWidth={true}
-                    style={{ marginTop: theme.spacing(2) }}
-                  />
-                  <Field
-                    component={TextField}
-                    name="confirmation"
-                    label={t_i18n('Confirmation')}
-                    type="password"
-                    fullWidth={true}
-                    style={{ marginTop: theme.spacing(2) }}
-                  />
-                  <Stack
-                    mt={3}
-                    direction="row"
-                    alignItems="center"
-                    justifyContent="flex-end"
-                  >
-                    <Button
-                      type="submit"
-                      disabled={isSubmitting || !isValid}
-                    >
-                      {t_i18n('Change your password')}
-                    </Button>
-                  </Stack>
+            <ForcePasswordChangeForm
+              onSuccess={handleSuccess}
+              renderPolicies={(password) => (
+                <Box sx={{ width: '100%', mt: 2 }}>
+                  <PasswordPolicies value={password} />
                 </Box>
-              </Form>
-            )}
-          </Formik>
+              )}
+            />
+          </Box>
         </Stack>
       </Stack>
       <Box
